Bounded Executor tests
