+++ src/main/java/com/jano7/executor/BoundedExecutor.java
@@ -43,7 +43,6 @@ public final class BoundedExecutor implements DrainableExecutor {
 
     private boolean drained = false;
 
-    // TODO test REJECT
     public BoundedExecutor(int maxTasks, BoundedStrategy onTasksExceeded, Executor underlyingExecutor) {
         this.maxTasks = maxTasks;
         this.semaphore = new Semaphore(maxTasks);
